Fix All GUI Warnings and Added the display system

+++ gui/source/Systems.cpp
@@ -9,7 +9,7 @@ sparse_array<Position> positions,
 sparse_array<Speed> speed, 
 sparse_array<Direction> dir)
 {
-    for (int ind = 0; ind < positions.size(); ind++) {
+    for (size_t ind = 0; ind < positions.size(); ind++) {
         std::cout << "pos : " << positions[ind]->pos_X << ", " << positions[ind]->pos_Y << "\nspeed : " << speed[ind]->speed << "\ndir : " << dir[ind]->dir_X << ", " << dir[ind]->dir_Y << std::endl;
         float magnitude = std::sqrt((dir[ind]->dir_X * dir[ind]->dir_X) + (dir[ind]->dir_Y * dir[ind]->dir_Y));
         std::cout << "magnitude " << magnitude << std::endl;
@@ -26,7 +26,7 @@ sparse_array<Position> positions,
 sparse_array<Size> size, 
 sparse_array<Sprite> sprite)
 {
-    for (int ind = 0; ind < sprite.size(); ind++) {
+    for (size_t ind = 0; ind < sprite.size(); ind++) {
         ClearBackground(BLACK);
         BeginDrawing();
             DrawTexture(sprite[ind]->sprite, positions[ind]->pos_X, positions[ind]->pos_Y, WHITE);
